Changes integer x and y coordinates to doubles
Changes the x and y coordinates as well as the speed vectors in the
Entity class to Point2D.Doubles for higher precision.

Organised imports as well.

+++ com/moyamo/bfc/entities/Projectile.java
@@ -1,19 +1,15 @@
 package com.moyamo.bfc.entities;
 
-import javax.xml.crypto.dsig.spec.XSLTTransformParameterSpec;
-
-import com.moyamo.bfc.debug.Out;
-
 public abstract class Projectile extends Entity implements IMovable{
 
-	public Projectile(int x, int y, int width, int height, int direction) {
+	public Projectile(double x, double y, int width, int height, int direction) {
 		super(x, y, width, height, direction);
 	}
 
 	@Override
 	public void move(long timeDiff) {
-		setX((int)(getX() + getXSpeed() * timeDiff/1000.0));
-		setY((int)(getY() + getYSpeed() * timeDiff/1000.0));
+		setX(getX() + getXSpeed() * timeDiff/1000.0);
+		setY(getY() + getYSpeed() * timeDiff/1000.0);
 	}
 	
 }
